ref(producers): extract service and ingress producers
